+++ app/src/main/java/com/example/associate/training/MainActivity.kt
@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.associate.training.amphibian.AmphibianActivity
-import com.example.associate.training.animation.AnimationActivity
+import com.example.associate.training.media.MediaActivity
 import com.example.associate.training.async.ThreadActivity
 import com.example.associate.training.busschedule.BusScheduleActivity
 import com.example.associate.training.canvas.CanvasActivity
@@ -38,7 +38,7 @@ open class MainActivity : AppCompatActivity() {
 
     open fun getList(): List<Entry> {
         return listOf(
-            Entry("AnimationActivity", AnimationActivity::class.java),
+            Entry("AnimationActivity", MediaActivity::class.java),
             Entry("KoinActivity", KoinActivity::class.java),
             Entry("ViewsActivity", ViewsActivity::class.java),
             Entry("MiscActivity", MiscActivity::class.java),
